Agent/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
schedule==1.2.1
ipfshttpclient==0.8.0
textract==1.6.5
numpy==1.24.3
pandas==2.0.3
networkx==3.1
matplotlib==3.7.2
plotly==5.15.0
geopandas==0.13.2
python-docx==0.8.11
pypdf==3.15.1